Allow creating project in existing directory
Currently, if you want to create a project in an existing (or the current) directory, `fs` will throw an error and exit, as the cli is trying to create a directory that already exists. The changes I've made allow for you to run `npx sandstone-cli create .` to create a sandstone project in your current directory, for example.